+++ packages/panda/src/theme/recipes/slider.ts
@@ -6,6 +6,9 @@ export const slider = defineSlotRecipe({
   slots: sliderAnatomy.keys(),
   base: {
     root: {
+      display: 'flex',
+      flexDirection: 'column',
+      gap: '1',
       width: 'full',
     },
     control: {
@@ -16,11 +19,13 @@ export const slider = defineSlotRecipe({
     track: {
       backgroundColor: 'bg.muted',
       borderRadius: 'l2',
+      // backgroundColor: 'bg.emphasized',
+      // borderRadius: 'full',
+      // overflow: 'hidden',
       flex: '1',
     },
     range: {
       background: 'accent.default',
-      borderRadius: 'l2',
     },
     thumb: {
       background: 'bg.default',
@@ -29,10 +34,35 @@ export const slider = defineSlotRecipe({
       borderWidth: '2px',
       boxShadow: 'sm',
       outline: 'none',
+      zIndex: '1',
     },
     label: {
       color: 'fg.default',
       fontWeight: 'medium',
+    },
+    markerGroup: {
+      mt: '-1',
+    },
+    marker: {
+      '--before-background': {
+        base: 'colors.white',
+        _dark: 'colors.accent.fg',
+      },
+      color: 'fg.muted',
+      _before: {
+        background: 'white',
+        borderRadius: 'full',
+        content: "''",
+        display: 'block',
+        left: '50%',
+        position: 'relative',
+        transform: 'translateX(-50%)',
+      },
+      _underValue: {
+        _before: {
+          background: 'var(--before-background)',
+        },
+      },
     },
   },
   defaultVariants: {
@@ -40,9 +70,35 @@ export const slider = defineSlotRecipe({
   },
   variants: {
     size: {
+      sm: {
+        control: {
+          height: '4',
+        },
+        range: {
+          height: '1.5',
+        },
+        track: {
+          height: '1.5',
+        },
+        thumb: {
+          height: '4',
+          width: '4',
+        },
+        marker: {
+          _before: {
+            height: '1',
+            top: '-2.5',
+            width: '1',
+          },
+          textStyle: 'sm',
+        },
+        label: {
+          textStyle: 'sm',
+        },
+      },
       md: {
         control: {
-          py: '2',
+          height: '5',
         },
         range: {
           height: '2',
@@ -51,15 +107,45 @@ export const slider = defineSlotRecipe({
           height: '2',
         },
         thumb: {
-          height: '6',
-          width: '6',
+          height: '5',
+          width: '5',
         },
         marker: {
-          mt: '2',
+          _before: {
+            height: '1',
+            top: '-3',
+            width: '1',
+          },
           textStyle: 'sm',
         },
         label: {
           textStyle: 'sm',
+        },
+      },
+      lg: {
+        control: {
+          height: '6',
+        },
+        range: {
+          height: '2.5',
+        },
+        track: {
+          height: '2.5',
+        },
+        thumb: {
+          height: '6',
+          width: '6',
+        },
+        marker: {
+          _before: {
+            height: '1.5',
+            top: '-15px',
+            width: '1.5',
+          },
+          textStyle: 'md',
+        },
+        label: {
+          textStyle: 'md',
         },
       },
     },