+++ website/src/components/copy-to-clipboard-button.tsx
@@ -33,6 +33,7 @@ export const CopyToClipboardButton = (props: Props) => {
           variant="ghost"
           size="sm"
           aria-label="Copy code to clipboard"
+          display={{ base: 'none', sm: 'inline-flex' }}
           onClick={handleClick}
           {...styles}
         >