+++ website/src/components/live-preview.tsx
@@ -1,10 +1,11 @@
-import { InfoIcon } from 'lucide-react'
+import { ExternalLinkIcon, InfoIcon } from 'lucide-react'
 import { generateElement } from 'react-live'
 import { Box, Flex, HStack, Stack } from 'styled-system/jsx'
 import * as components from '~/components/ui'
+import { Demo } from './demos/checkbox.demo'
 
-const icons = { InfoIcon }
-const scope = { ...components, ...icons, Stack, HStack, Box, Flex }
+const icons = { InfoIcon, ExternalLinkIcon }
+const scope = { ...components, ...icons, Stack, HStack, Box, Flex, Checkbox: Demo }
 
 type Props = {
   code: string